+++ eventregistry/tests/TestERTopicPage.py
@@ -37,6 +37,31 @@ class TestTopicPage(DataValidator):
                 uriSet.add(event.get("uri"))
 
 
+    def testCreateTopicPage(self):
+        topic = TopicPage(self.er)
+        appleUri = self.er.getConceptUri("apple")
+        msoftUri = self.er.getConceptUri("microsoft")
+        iphoneUri = self.er.getConceptUri("iphone")
+        businessUri = self.er.getCategoryUri("business")
+        topic.addConcept(appleUri, 50, required = False)
+        topic.addConcept(msoftUri, 50, required = True)
+        topic.addConcept(iphoneUri, 50, excluded = True)
+        topic.addCategory(businessUri, 50, required=True)
+        for page in range(1, 10):
+            res = topic.getArticles(page = page, returnInfo = ReturnInfo(articleInfo=ArticleInfoFlags(concepts=True, categories=True, maxConceptsPerType=100)))
+            for art in res.get("articles").get("results"):
+                foundConcept = False
+                foundCategory = False
+                for conceptObj in art.get("concepts", []):
+                    assert iphoneUri != conceptObj["uri"], "Found iphone in the article"
+                    if msoftUri == conceptObj["uri"]:
+                        foundConcept = True
+                for categoryObj in art.get("categories", []):
+                    if categoryObj["uri"].startswith(businessUri):
+                        foundCategory = True
+                assert foundConcept, "Article did not have a required concept"
+                assert foundCategory, "Article did not have a required category"
+
 
 if __name__ == "__main__":
     suite = unittest.TestLoader().loadTestsFromTestCase(TestTopicPage)